% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.model.sa2-opts.R
\name{covid19.model.sa2-opts}
\alias{covid19.model.sa2-opts}
\title{covid19.model.sa2 package options}
\description{
covid19.model.sa2 package options
}
\section{Options}{


The following package options are available and set to the
given default when loaded.
\describe{
\item{\code{covid19.model.sa2_dataEnv = new.env()}}{Environment into which data will be saved/accessed.}
\item{\code{covid19.model.sa2_nThread = 1L}}{Number of threads to use.}
\item{\code{covid19.model.sa2.fst2_progress = interactive()}}{Number of threads to use.}
}
}

