% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.model.sa2-package.R
\docType{package}
\name{covid19.model.sa2-package}
\alias{covid19.model.sa2}
\alias{covid19.model.sa2-package}
\title{covid19.model.sa2: Model Policy and Epidemiological Assumptions of the COVID-19 Pandemic}
\description{
A dynamic microsimulation model of Australia, taking policy changes and
    epidemiological assumptions about transmissions.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Will Mackey
}

Other contributors:
\itemize{
  \item Daniel Lemire (Adapted a header of his to generate the Lehmer PRNG) [contributor]
}

}
\keyword{internal}
