% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sys.R
\name{read_sys}
\alias{read_sys}
\title{Read package system file}
\usage{
read_sys(
  file.fst,
  columns = NULL,
  pattern = NULL,
  fst2_progress = getOption("covid19.model.sa2.fst2_progress"),
  ...
)
}
\arguments{
\item{file.fst}{\describe{
  \item{\code{character(1)}}{
    A string representing the path to the file.
    For convenience, can be of the form \code{<ENV VAR>:/<branch>} where
    \code{<ENV VAR>} is the name of an environment variable set to a
    path.
}}}

\item{columns}{\describe{\item{\code{character(n)}}{
As in \code{\link[fst:write_fst]{read_fst}}, a subset of columns
from \code{file.fst}. It differs from \code{\link[fst:write_fst]{read_fst}}
by only taking the intersection of \code{columns} and the columns
actually available.
}}}

\item{pattern}{\describe{\item{\code{NULL; character(1)}}{
A regular expression,
another way to select columns from
\code{file.fst}.

If \code{NULL}, the default, no additional columns are selected.
}}}

\item{fst2_progress}{Whether to write progress to the console.}

\item{...}{Arguments passed to \code{read_fst}.}
}
\description{
Read package system file
}
\examples{
read_sys("time_series_cases.fst")
read_sys("time_series_cases.fst", pattern = "^[A-Z]+$", columns = "Date")


}
