% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-platform.R
\name{run_covid19_example}
\alias{run_covid19_example}
\title{Platform utilities}
\usage{
run_covid19_example(pkgs = c("tibble", "testthat", "covr"))
}
\arguments{
\item{pkgs}{Packages required. Defaults to suggested packages.}
}
\value{
\code{TRUE} on 64-bit environments with suggested packages.
}
\description{
Various utilities to determine the suitability of
the environment for running, such as 64-bit and sufficient memory.
}
