% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_epipars.R
\name{set_epipars}
\alias{set_epipars}
\title{Set epidemiological parameters}
\usage{
set_epipars(
  incubation_distribution = c("pois", "lnorm", "dirac"),
  incubation_mean = 5,
  incubation_sigma = 0.44,
  illness_distribution = c("pois", "lnorm", "dirac"),
  illness_mean = 15,
  illness_sigma = 1,
  r_distribution = c("cauchy", "lnorm", "pois", "dirac"),
  r_location = 2,
  r_scale = 1,
  p_asympto = 0.48,
  p_critical = 0.02,
  p_death = 0.01
)
}
\arguments{
\item{incubation_distribution}{One of `"pois"`, `"lnorm"`, and `"dirac"`,
whether the incubation period is Poisson, log-normal, or constant.}

\item{incubation_mean}{The intended average of the distribution, must be
positive, and a whole number for `"pois"` and `"diract"`.}

\item{incubation_sigma}{A measure of the spread of the distribution.}

\item{illness_distribution, illness_mean, illness_sigma}{As for \code{incubation} above.}

\item{r_distribution}{The distribution of the number of infections from each
infected person, one of `"cauchy"`, `"lnorm"`, `"pois"`, or `"dirac"`.}

\item{r_location, r_scale}{Parameters for \code{r_distribution}.}

\item{p_asympto}{A number in \code{[0, 1]}, the proportion of cases that
are asymptomatic.}

\item{p_critical}{A number in \code{[0, 1]}, the proportion of
\strong{symptomatic} patients that require ICU.}

\item{p_death}{A number in \code{[0, 1]}, the proportion of
\strong{critical cases} that die.}
}
\value{
A list of the components, plus an entry \code{CHECKED} having the
value \code{TRUE}.
}
\description{
Used to supply epidemiological parameters to the main
simulate function, with defaults.
}
