% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_initial.R
\name{set_initial_by_state}
\alias{set_initial_by_state}
\title{Set the status among a population}
\usage{
set_initial_by_state(
  state_id,
  dead = NULL,
  healed = NULL,
  active = NULL,
  critical = NULL,
  cases = NULL,
  cases_by_state = "time_series_cases.fst",
  deaths_by_state = "time_series_deaths.fst",
  recovered_by_state = "time_series_recovered.fst",
  asympto = 0.48,
  p_critical = 0.03,
  first_yday = NULL,
  p_isol = NULL,
  .population = NULL
)
}
\arguments{
\item{state_id}{An integer vector of states, the match of
 "NSW", "VIC", "QLD", "WA", "SA", "TAS", "ACT", "NT".
If \code{dead = NULL} then \code{state_id} must be length-one.}

\item{dead}{The number of patients who have died.}

\item{healed}{The number of patients who have recovered.}

\item{active}{The number of patients who have tested positive
and have not recovered, died, or been admitted to intensive care.}

\item{critical}{The number of patients in intensive care.}

\item{cases}{Optionally, the number of cumulative cases. That is,
\code{dead + healed + active + critical}.}

\item{cases_by_state}{A table of total cases (cumulative) by date,
ordered by date, and having columns matching the states in \code{state_id}.
Alternatively a string to be read in by \code{\link{read_sys}}, the path
to a file with this data.}

\item{deaths_by_state, recovered_by_state}{As for \code{cases_by_state}.}

\item{asympto}{The proportion of cases who are asymptomatic.
\strong{Note: active cases are assumed to be all symptomatic. Thus,
if the number of active cases is 100 and }\code{asympto = 0.75}
\strong{ the number of total cases is 400.}}

\item{p_critical}{The proportion of symptomatic cases who are critical.}

\item{first_yday}{An integer, the \code{\link[data.table:IDateTime]{yday}} to set
as the initial day.}

\item{p_isol}{The proportion of cases in isolation.}

\item{.population}{The size of the output, the population. By default,
set to the length of \code{state_id}.}
}
\value{
An integer vector of length \code{.population} with values
as given in the status table \code{\link{simulate_sa2}}. The statuses
are randomly provided throughout the vector in the distribution of
\code{dead, healed, active, critical}.
}
\description{
Set the status among a population
}
