% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_policypars.R
\name{set_policypars}
\alias{set_policypars}
\title{Set policy parameters}
\usage{
set_policypars(
  supermarkets_open = TRUE,
  schools_open = FALSE,
  only_Year12 = FALSE,
  do_contact_tracing = TRUE,
  tests_by_state = NULL,
  max_persons_per_event = 5L,
  max_persons_per_supermarket = 200L
)
}
\arguments{
\item{supermarkets_open}{\code{TRUE|FALSE}. Should supermarkets remain open?}

\item{schools_open}{\code{FALSE|TRUE}. Should schools remain open?}

\item{only_Year12}{\code{FALSE|TRUE}. If schools open, should they be
restricted to Year 12 students only? No effect if \code{schools_open = FALSE}.}

\item{do_contact_tracing}{\code{TRUE|FALSE}. Should contact tracing occur?
If \code{FALSE} households are not isolated if tested.}

\item{tests_by_state}{\code{integer(10)} The number of tests per day that states
perform. First entry is the total tests available across Australia.
If any entry is negative,
or \code{NA}, the most recent day's number of tests performed are
used.}
}
\value{
A list of the components.
}
\description{
Used to supply policy parameters to the main
\code{\link{simulate_sa2}} function,
with defaults.
}
