% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{states}
\alias{states}
\title{States as coded}
\usage{
states()
}
\description{
State names are used throughout the package and are encoded
when the simulation is run. This is a convenience function that simply
returns a character vector of the state abbreviations recognized
by other functions.

N.B. The vector is intended to be zero-indexed. So AUS is the
zeroth element of \code{states}.
}
