% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pappubahry.R
\name{update_pappubahry}
\alias{update_pappubahry}
\title{Update the time series data from pappubahry/COVID19}
\usage{
update_pappubahry()
}
\value{
A named logical vector, indicating success for the given file.

Invoked for its side-effect: changing the internal \code{fst} files
giving the time series of COVID-19 cases from
\url{https://github.com/HughParsonage/pappubahry-AU_COVID19}.

Note that this changes the package's files: any original files will be
overwritten.
}
\description{
Update the time series data from pappubahry/COVID19
}
