% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.model.sa2-package.R
\docType{package}
\name{covid19.model.sa2-package}
\alias{covid19.model.sa2}
\alias{covid19.model.sa2-package}
\title{covid19.model.sa2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
