% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_epipars.R
\name{set_epipars}
\alias{set_epipars}
\title{Set epidemiological parameters}
\usage{
set_epipars(
  incubation_distribution = c("pois", "lnorm", "dirac"),
  incubation_mean = 5,
  incubation_sigma = 0.44,
  illness_distribution = c("pois", "lnorm", "dirac"),
  illness_mean = 15,
  illness_sigma = 1,
  r_distribution = c("cauchy", "lnorm", "pois", "dirac"),
  r_location = 2/5,
  r_scale = 1/5,
  r_schools_distribution = r_distribution,
  r_schools_location = r_location * 0.25,
  r_schools_scale = r_scale,
  r_supermarket_location = r_location,
  r_supermarket_scale = r_scale,
  r_work_location = r_location,
  a_workplace_rate = 0.07,
  resistance_threshold = 400L,
  p_asympto = 0.48,
  p_critical = 0.02,
  p_death = 0.01
)
}
\arguments{
\item{incubation_distribution}{One of `"pois"`, `"lnorm"`, and `"dirac"`,
whether the incubation period is Poisson, log-normal, or constant.}

\item{incubation_mean}{The intended average of the distribution, must be
positive, and a whole number for `"pois"` and `"diract"`.}

\item{incubation_sigma}{A measure of the spread of the distribution.}

\item{illness_distribution, illness_mean, illness_sigma}{As for \code{incubation} above.}

\item{r_distribution}{The distribution of the number of infections from each
infected person, one of `"cauchy"`, `"lnorm"`, `"pois"`, or `"dirac"`.}

\item{r_location, r_scale}{Parameters for \code{r_distribution}. The number of
infections \strong{per day}.

If \code{r_distribution = "dirac"} there is no distribution;
the infection will be deterministic. If an integer is provided, all infectious
persons infect the same number of people. Can also be provided in the form
\code{a/b} where \code{a} and {b} are whole numbers, in which case each
infectious person will infect \code{a} individuals every \code{b} days
(precisely, not on average).}

\item{r_schools_distribution, r_schools_location, r_schools_scale}{Variables
particular for schools.}

\item{r_supermarket_location, r_supermarket_scale}{Variables particular
for supermarket.}

\item{r_work_location}{Parameters for work infection distribution.}

\item{a_workplace_rate}{\code{double(1)} in \code{[0, 1]}, the (average)
proportion of colleagues that will be infected in any day.}

\item{resistance_threshold}{An integer in \code{[0, 1000]}, the resistance
required to not be infected. A value of 0
means no-one will be infected; a value of 1000 means everyone will.}

\item{p_asympto}{A number in \code{[0, 1]}, the proportion of cases that
are asymptomatic.}

\item{p_critical}{A number in \code{[0, 1]}, the proportion of
\strong{symptomatic} patients that require ICU.}

\item{p_death}{A number in \code{[0, 1]}, the proportion of
\strong{critical cases} that die.}
}
\value{
A list of the components, plus an entry \code{CHECKED} having the
value \code{TRUE}.
}
\description{
Used to supply epidemiological parameters to the main
simulate function, with defaults.
}
