% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_epipars.R
\name{set_epipars}
\alias{set_epipars}
\title{Set epidemiological parameters}
\source{
\url{https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(20)30287-5/fulltext}
for the attack rates.

See also AceMOD

\describe{
\item{a_household_rate}{12.1 to 18.2 or (9.1 to 13.8 assuming missing tests were uninfected).}
}
}
\usage{
set_epipars(
  incubation_distribution = c("pois", "lnorm", "dirac", "cauchy"),
  incubation_mean = 5,
  incubation_sigma = 0.44,
  illness_distribution = c("pois", "lnorm", "dirac"),
  illness_mean = 15,
  illness_sigma = 1,
  a_workplace_rate = 0.07,
  a_household_rate = 0.15,
  a_schools_rate = 0.07,
  q_workplace = 0.01,
  q_household = 0.05,
  q_school = 1/3000,
  q_school_grade = 1/500,
  q_supermarket = 1/500,
  q_places = 1/500,
  q_major_event = 1/5000,
  resistance_threshold = 400L,
  p_asympto = 0.48,
  p_critical = 0.02,
  p_death = 0.01,
  p_visit_major_event = 1/52,
  supermarket_beta_shape1 = 3,
  supermarket_beta_shape2 = 1
)
}
\arguments{
\item{incubation_distribution}{One of `"pois"`, `"lnorm"`, and `"dirac"`,
whether the incubation period is Poisson, log-normal, or constant.}

\item{incubation_mean}{The intended average of the distribution, must be
positive, and a whole number for `"pois"` and `"dirac"`.}

\item{incubation_sigma}{A measure of the spread of the distribution.}

\item{illness_distribution, illness_mean, illness_sigma}{As for \code{incubation} above.}

\item{a_workplace_rate}{\code{double(1)} in \code{[0, 1]}, the proportion
of workplaces in which transmission between colleagues is possible.}

\item{a_household_rate}{\code{double(1)} in \code{[0, 1]}, the
proportion of households in which household transmission is possible.}

\item{a_schools_rate}{\code{double(1)} in \code{[0, 1]}, the proportion of
schools in which tranmission between students is possible.}

\item{q_household, q_school, q_school_grade, q_workplace}{Daily transmission probability among household members / students of the same school / students of the same school and same same
grade / workers of the same workplace.}

\item{q_supermarket, q_places, q_major_event}{As above, for supermarket and places.}

\item{resistance_threshold}{An integer in \code{[0, 1000]}, the resistance
required to not be infected. A value of 0
means no-one will be infected; a value of 1000 means everyone will.}

\item{p_asympto}{A number in \code{[0, 1]}, the proportion of cases that
are asymptomatic.}

\item{p_critical}{A number in \code{[0, 1]}, the proportion of
\strong{symptomatic} patients that require ICU.}

\item{p_death}{A number in \code{[0, 1]}, the proportion of
\strong{critical cases} that die.}

\item{p_visit_major_event}{Probability a person attends a major event on a given day.}

\item{supermarket_beta_shape1, supermarket_beta_shape2}{Parameters for the
\code{rbeta} function that determines the distribution of visit frequency
to supermarkets.}
}
\value{
A list of the components, plus an entry \code{CHECKED} having the
value \code{TRUE}.
}
\description{
Used to supply epidemiological parameters to the main
simulate function, with defaults.

Note that all parameters are taken together. In particular, attack probabilities,
transmission probabilities, and reproduction parameters affect the overall
transmission multiplicatively .
}
