% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_policypars.R
\name{set_policypars}
\alias{set_policypars}
\title{Set policy parameters}
\usage{
set_policypars(
  supermarkets_open = TRUE,
  schools_open = FALSE,
  only_Year12 = FALSE,
  school_days_per_wk = NULL,
  do_contact_tracing = TRUE,
  contact_tracing_days_before_test = 0L,
  contact_tracing_days_until_result = 3L,
  contact_tracing_only_sympto = TRUE,
  tests_by_state = NULL,
  max_persons_per_event = 5L,
  max_persons_per_supermarket = 200L
)
}
\arguments{
\item{supermarkets_open}{\code{TRUE|FALSE}. Should supermarkets remain open?}

\item{schools_open}{\code{FALSE|TRUE}. Should schools remain open?}

\item{only_Year12}{\code{FALSE|TRUE}. If schools open, should they be
restricted to Year 12 students only? No effect if \code{schools_open = FALSE}.}

\item{do_contact_tracing}{\code{TRUE|FALSE}. Should contact tracing occur?
If \code{FALSE} households are not isolated if tested.}

\item{contact_tracing_days_before_test}{The number of days following
the end of the incubation period
before the person gets tested.}

\item{contact_tracing_days_until_result}{The number of days between a test
and the result being known.}

\item{contact_tracing_only_sympto}{\code{TRUE|FALSE} Is contact tracing only
applied to symptomatic cases?}

\item{tests_by_state}{\code{integer(10)} The number of tests per day that states
perform. First entry is the total tests available across Australia.
If any entry is negative,
or \code{NA}, the most recent day's number of tests performed are
used.}

\item{max_persons_per_event, max_persons_per_supermarket}{Not yet used.}

\item{schools_days_per_wk}{Specifies how many days a week pupils attend
school. By default, students attend full-time. Only applied after
\code{schools_open} and \code{only_Year12}. In particular, has no
effect if \code{schools_open = FALSE} and if \code{only_Year12 = TRUE} then
all other students attend 0 times per week.

If a single number, applies to all states. Otherwise must be a
named list or vector.

If a named vector, names must be state names.
(See \code{\link{states}}.) Omitted states take the value 5 (i.e.
full-time).

If a named list, each element must be a length-21 vector which
specifies the number of times per week students of each age 0-20
attend in the given state (with the last
element being the attendance per week of teachers).'}
}
\value{
A list of the components.
}
\description{
Used to supply policy parameters to the main
\code{\link{simulate_sa2}} function,
with defaults.
}
