% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_sa2}
\alias{simulate_sa2}
\title{Simulate this}
\usage{
simulate_sa2(
  days_to_simulate = 300,
  InitialStatus = list(dead = 63, healed = 3747, active = 2592, critical = 66),
  CauchyM = integer(0),
  EpiPars = list(),
  .first_day = yday(Sys.Date()),
  .population = 2.5e+07,
  verbose_timer = TRUE
)
}
\description{
Simulate this
}
