% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_sa2}
\alias{simulate_sa2}
\title{Simulate the spread of COVID-19 through Australia}
\usage{
simulate_sa2(
  days_to_simulate = 5,
  PolicyPars = set_policypars(),
  EpiPars = set_epipars(),
  InitialStatus = list(dead = 71, healed = 4685, active = 1840, critical = 49),
  .first_day = NULL,
  showProgress = 1L,
  by_state = TRUE,
  dataEnv = getOption("covid19.model.sa2_dataEnv", new.env()),
  use_dataEnv = getOption("covid19.model.sa2_useDataEnv", FALSE),
  nThread = getOption("covid19.model.sa2_nThread", 1L),
  myaus = NULL
)
}
\arguments{
\item{days_to_simulate}{\describe{
\item{\code{integer(1)}}{The number of days to simulate.}
}}

\item{PolicyPars}{\describe{
\item{\code{list()}}{A list describing the policy parameters.
(\code{list(schools_open = TRUE|FALSE, only_Year12 = TRUE|FALSE)}
are supported but no others. An empty list corresponds to current settings.)}
}}

\item{EpiPars}{\describe{
\item{\code{list()}}{A list of epidemiological parameters generated by \code{\link{set_epipars}}.}
}}

\item{InitialStatus}{\describe{
\item{\code{list(4)}}{A list of four named components giving the
number of dead, healed, active, and critical on the day zero.}
}}

\item{.first_day}{\describe{
\item{\code{integer(1)}}{The first day to model.
If \code{NULL}, the default, the day of the year of the most recent date
in records from Australia. Used to predict the outcomes of those presently
ill with COVID-19 based on imputed duration of infection.}
}}

\item{showProgress}{Should progress times and bars be displayed?
By default, 1 which displays bars unless it is likely to not be
desired (for example it won't be displayed when running tests or
in non-interactive sessions). Set to 0 to never display; 2 to always
display. Set to 3 to also show internal data access.}

\item{by_state}{Set the initial status by state.}

\item{dataEnv}{An environment into which the data will be saved and
retrieved. Used to save time reading after the first time.}

\item{use_dataEnv}{\code{FALSE|TRUE} Whether or not to save the
underlying Australia table to \code{dataEnv} (and retrieve it on
subsequent uses).}

\item{nThread}{\describe{
\item{\code{integer(1)}}{Number of threads to use in the C++ code.}
}}

\item{myaus}{A prepared \code{data.table} with a column \code{status}
that has the desired distribution (including perhaps over SA2). This
formal is experimental and may be removed without notice.}
}
\value{
A list of \code{days_to_simulate + 1} components. The first
component is the inital status of each individual and
subsequent components are the statuses of each day simulated.
}
\description{
Simulates the spread of COVID-19 given today's condition, and
the (imputed) distribution of indivdiuals among households,
schools, and other places, from day to day.
}
\details{
The \strong{Status} of an individual is\tabular{rl}{
   Status \tab Description \cr
   -2 \tab Dead \cr
   -1 \tab Healed \cr
   0 \tab Susceptible \cr
   1 \tab Infected, not showing symptoms \cr
   2 \tab Infected, showing symptoms \cr
   3 \tab Critical \cr
   32 \tab 0 but isolated \cr
   33 \tab 1 but isolated \cr
}
}
